requirements.txt
python-dotenv
supabase
pandas
polars
pyarrow
postgrest-py
connectorx